Les transferts marchent, mais y'a pas encore l'animation

+++ superflu-android/src/com/tacoid/superflu/entities/Entity.java
@@ -1,5 +1,5 @@
 package com.tacoid.superflu.entities;
 
 public interface Entity {
-	public void update(float delta);
+	public void update(int delta);
 }
